Move directive implementations into an own package
