+++ src/main/kotlin/net/theevilreaper/dartpoet/directive/impl/ExportDirective.kt
@@ -1,7 +1,7 @@
-package net.theevilreaper.dartpoet.directive
+package net.theevilreaper.dartpoet.directive.impl
 
 import net.theevilreaper.dartpoet.code.CodeWriter
-import net.theevilreaper.dartpoet.util.SEMICOLON
+import net.theevilreaper.dartpoet.directive.*
 
 /**
  * This implementation represents the export directive from dart.
@@ -13,17 +13,17 @@ import net.theevilreaper.dartpoet.util.SEMICOLON
  * @author theEvilReaper
  */
 class ExportDirective internal constructor(
-    private val path: String,
+    path: String,
     private val castType: CastType? = null,
     private val importCast: String? = null,
-) : BaseDirective(path) {
+) : BaseDirective(DirectiveType.EXPORT, path) {
 
     private val export = "export"
     private val invalidCastType = arrayOf(CastType.DEFERRED, CastType.AS)
 
     init {
-        if (castType != null && castType in invalidCastType) {
-            throw IllegalStateException("The following cast types are not allowed for an export directive: [${invalidCastType.joinToString()}]")
+        check(!(castType != null && castType in invalidCastType)) {
+            "The following cast types are not allowed for an export directive: [${invalidCastType.joinToString()}]"
         }
 
         if (importCast != null) {
@@ -40,14 +40,6 @@ class ExportDirective internal constructor(
      * @param writer the [CodeWriter] instance to append the directive
      */
     override fun write(writer: CodeWriter) {
-        val ensuredPath = path.ensureDartFileEnding()
-        writer.emit("$export·'")
-        writer.emit(ensuredPath)
-        writer.emit("'")
-
-        if (importCast != null && castType != null) {
-            writer.emit("·${castType.identifier} $importCast")
-        }
-        writer.emit(SEMICOLON)
+        DirectiveHelper.writeDirective(writer, this, importCast, castType)
     }
 }
